updating optimization AEP to include proper intialization of turbine Cp, Ct

+++ src/wakeexchange/GeneralWindFarmComponents.py
@@ -604,7 +604,7 @@ class SpacingComp(Component):
                        desc='y coordinates of turbines in wind dir. ref. frame')
 
         # Explicitly size output array
-        self.add_output('wtSeparationSquared', val=np.zeros(((nTurbines-1)*nTurbines/2.)),
+        self.add_output('wtSeparationSquared', val=np.zeros(int((nTurbines-1)*nTurbines/2)),
                         desc='spacing of all turbines in the wind farm')
 
     def solve_nonlinear(self, params, unknowns, resids):
@@ -613,7 +613,7 @@ class SpacingComp(Component):
         turbineX = params['turbineX']
         turbineY = params['turbineY']
         nTurbines = turbineX.size
-        separation_squared = np.zeros(((nTurbines-1)*nTurbines/2))
+        separation_squared = np.zeros(int((nTurbines-1)*nTurbines/2))
 
         k = 0
         for i in range(0, nTurbines):
@@ -735,8 +735,8 @@ class BoundaryComp(Component):
             unit_normals = params['boundaryNormals']
 
             # initialize array to hold distances from each point to each face
-            dfaceDistance_dx = np.zeros([self.nTurbines*self.nVertices, self.nTurbines])
-            dfaceDistance_dy = np.zeros([self.nTurbines*self.nVertices, self.nTurbines])
+            dfaceDistance_dx = np.zeros([int(self.nTurbines*self.nVertices), self.nTurbines])
+            dfaceDistance_dy = np.zeros([int(self.nTurbines*self.nVertices), self.nTurbines])
 
             for i in range(0, self.nTurbines):
                 # determine if point is inside or outside of each face, and distance from each face
